// src/app/maps/types/TiledLayers.ts
export enum TiledLayers {
  platforms = 'platforms/platforms',
  platformsColider = 'platforms/platforms-colider',
  platformsBottom = 'platforms/platforms-bottom',
  grass = 'environment/grass',
  environment = 'environment/environment',
  grass_top = 'environment/grass-top',
  fance = 'environment/fance',
}